fix: properly restore values matching multiple paths
**The issue**

Each path type (static, group, nested)
uses an **isolated** way of storing and restoring
values on the original object.

In case of a property that matches multiple
redaction paths, this isolation leads to loss
of information as it fails to restore the
original value on the original mutated object.

The issue can be reproduced with the following:

```js
const redact = fastRedact({
  paths: ['a', '*.b', 'x.b'],
  serialize: false
})
const o = {
  x: {
    a: 'a',
    b: 'b'
  }
}
```

In this example `x.b` would match both the static `x.b`
path and the wildcard `*.b` path leading in the original
object being restored as:

```
{
  x: {
    a: 'a',
    b: '[REDACTED]'
  }
}
```

**Resolution**

This fix tackles the problem by introducing a global "stash"
that is shared across the various (static, group, nested)
redaction types, protecting a redaction from ovewriting
an original value with a redacted value produced from
a previous redaction.