fix: properly restore values matching multiple paths
Each path type (static, group, nested)
uses an **isolated** way of storing and restoring
values on the original object.

In case of a property that matches multiple
redaction paths, this isolation leads to loss
of information as it fails to restore the
original value on the original mutated object.

The issue can be reproduced with the following:

```js
const redact = fastRedact({
  paths: ['a', '*.b', 'x.b'],
  serialize: false
})
const o = {
  x: {
    a: 'a',
    b: 'b'
  }
}
redact(o)
```

In this example `x.b` would match both the static `x.b`
path and the wildcard `*.b` path leading in the original
object being restored as:

```
{
  x: {
    a: 'a',
    b: '[REDACTED]'
  }
}
```

See https://github.com/davidmarkclements/fast-redact/issues/33

Restoration of the redacted object happens
backwards, in reverse order of redactions,
so that repeated redactions on the same object
property can be eventually rolled back to the
original value.

This way dynamic redactions are restored first,
starting from the last one working backwards and
followed by the static ones.